Remove sending transition
I needed to introduce a delay on the server or the client to make it play reliably. I think it could be caused by the .sending class being removed from the element before Vue has rendered it.

Getting this to work is too much effort and it might even look better without a transition.